Move logIfSlow to sqlClient.ts, breaking an import cycle
